Implemented proof-of-concept track import from DLNA server.
At the moment, we are using the Search function to retrieve all
music objects that do not have refID attribute (i.e., are not a
virtual resource). This functionality is probably missing from
many DLNA servers, but at the moment, it allows me to play music
from my rygel and minidlna shares.